Move compatibility modules into compat package.
