+++ distutils/sysconfig.py
@@ -16,7 +16,7 @@ import re
 import sys
 import sysconfig
 
-from . import py39compat
+from .compat import py39
 from ._functools import pass_none
 from .errors import DistutilsPlatformError
 
@@ -538,7 +538,7 @@ def get_config_vars(*args):
     global _config_vars
     if _config_vars is None:
         _config_vars = sysconfig.get_config_vars().copy()
-        py39compat.add_ext_suffix(_config_vars)
+        py39.add_ext_suffix(_config_vars)
 
     return [_config_vars.get(name) for name in args] if args else _config_vars
